[Truffle] Don't modify bin/ruby if we're just going to symlink it to the same thing.

+++ tool/jt.rb
@@ -166,8 +166,11 @@ module Utilities
   def self.find_jruby_bin_dir
     # Make sure bin/ruby points to the right launcher
     ruby_symlink = "#{JRUBY_DIR}/bin/ruby"
-    FileUtils.rm_f ruby_symlink
-    File.symlink Utilities.find_jruby, ruby_symlink
+    jruby_bin = Utilities.find_jruby
+    if File.symlink?(ruby_symlink) && File.expand_path(File.readlink(ruby_symlink), File.dirname(ruby_symlink)) != jruby_bin
+      FileUtils.rm_f ruby_symlink
+      File.symlink jruby_bin, ruby_symlink
+    end
 
     if ENV['RUBY_BIN']
       File.dirname(ENV['RUBY_BIN'])
